% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\alias{construct_}
\title{CONSTRUCT}
\usage{
construct(.data, ...)

construct_(.data, ..., .dots)
}
\arguments{
\item{.data}{input. This can be JSON input, or an object of class
\code{sparql_dsl} that has JSON and query params combined, which is passed
from function to function when using the sparqldsl DSL.}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
CONSTRUCT
}
\examples{
\dontrun{
### CONSTRUCT { ?s ?p ?o } WHERE { ?s ?p ?o } LIMIT 10
sparql_dsl("http://dbpedia.org/sparql") \%>\%
 construct(s, p, o) \%>\%
 where(s, p, o) \%>\%
 limit(10)
}
}
