% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional.R
\name{optional}
\alias{optional}
\alias{optional_}
\title{OPTIONAL}
\usage{
optional(.data, ...)

optional_(.data, ..., .dots)
}
\arguments{
\item{.data}{input. This can be JSON input, or an object of class
\code{sparql_dsl} that has JSON and query params combined, which is passed
from function to function when using the sparqldsl DSL.}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
OPTIONAL
}
\examples{
\dontrun{
### OPTIONAL { ?s ?p ?o }
conn <- sparql_dsl("http://dbpedia.org/sparql")
conn \%>\% optional(x, y) \%>\% inspect
sparql_dsl("http://dbpedia.org/sparql") \%>\% where(s, p, o, )
}
}
