% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{prefix}
\alias{prefix}
\alias{prefix_}
\title{PREFIX}
\usage{
prefix(.data, ...)

prefix_(.data, ..., .dots)
}
\arguments{
\item{.data}{input. This can be JSON input, or an object of class
\code{sparql_dsl} that has JSON and query params combined, which is passed
from function to function when using the sparqldsl DSL.}

\item{...}{(logical) named prefixes}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
PREFIX
}
\examples{
\dontrun{
### PREFIX cito: <http://purl.org/spar/cito/>
x <- sparql_dsl("http://dbpedia.org/sparql")
x \%>\% prefix(cito = "http://purl.org/spar/cito/") \%>\% inspect
x \%>\% select() \%>\% prefix(dcterms = "http://purl.org/dc/terms/") \%>\% inspect
x \%>\% prefix(cito = "http://purl.org/spar/cito/", dcterms = "http://purl.org/dc/terms/") \%>\% inspect
}
}
