% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{select_}
\title{SELECT}
\usage{
select(.data, ..., distinct = FALSE)

select_(.data, ..., .dots, distinct = FALSE)
}
\arguments{
\item{.data}{input. This can be JSON input, or an object of class
\code{sparql_dsl} that has JSON and query params combined, which is passed
from function to function when using the sparqldsl DSL.}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
SELECT
}
\examples{
\dontrun{
### SELECT * WHERE { ?s ?p ?o } OFFSET 100 LIMIT 10
### query = sparql.select.where([:s, :p, :o]).offset(100).limit(10)
sparql_dsl("http://dbpedia.org/sparql") \%>\%
 select() \%>\%
 where(s, p, o) \%>\%
 limit(10)
}
}

