% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_dsl.R
\name{sparql_dsl}
\alias{sparql_dsl}
\title{SPARQL dsl}
\usage{
sparql_dsl(url = NULL)
}
\arguments{
\item{url}{A SPARQL url}

\item{x}{for print}

\item{...}{ignored}
}
\description{
SPARQL dsl
}
\examples{
\dontrun{
sparql_dsl("http://dbpedia.org/sparql")

# ask
### should separate where from within ask, so ask() \%>\% where()
sparql_dsl("http://dbpedia.org/sparql") \%>\% ask(s, p, o)

# select
### SELECT * WHERE { ?s ?p ?o } OFFSET 100 LIMIT 10
### query = sparql.select.where([:s, :p, :o]).offset(100).limit(10)
sparql_dsl("http://dbpedia.org/sparql") \%>\%
 select() \%>\%
 where(s, p, o) \%>\%
 limit(10)

# construct
### CONSTRUCT { ?s ?p ?o } WHERE { ?s ?p ?o } LIMIT 10
sparql_dsl("http://dbpedia.org/sparql") \%>\%
 construct(s, p, o) \%>\%
 where(s, p, o) \%>\%
 limit(10) \%>\%
 inspect
}
}

