% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_exec.R
\name{sparql_exec}
\alias{sparql_exec}
\title{Execute sparql request}
\usage{
sparql_exec(x, ...)
}
\arguments{
\item{...}{Further arguments passed on to \code{\link[httr]{GET}}}

\item{url}{a SPARQL server url}

\item{query}{(character) A SPARQL query}
}
\description{
Execute sparql request
}
\examples{
\dontrun{
sparql_exec(spqurl())

out <- sparql_exec(spqurl(), query)
head(out)
}
}

