% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_query.R
\name{sparql_query}
\alias{sparql_query}
\title{SPARQL query, end around the DSL}
\usage{
sparql_query(.data, x, ...)
}
\arguments{
\item{.data}{a \code{sparql_dsl} class object, result of a call to
\code{\link{sparql_dsl}}}

\item{x}{A SPARQL query}

\item{...}{Further arguments passed on to \code{\link[httr]{GET}}}
}
\description{
SPARQL query, end around the DSL
}
\examples{
\dontrun{
sparql_dsl("http://dbpedia.org/sparql") \%>\%
 sparql_query("SELECT * WHERE { ?s ?p ?o } OFFSET 100 LIMIT 10")
}
}

