% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparqldsl-package.R
\docType{package}
\name{sparqldsl-package}
\alias{sparqldsl-package}
\alias{sparqldsl}
\title{SPARQL DSL Client}
\description{
SPARQL DSL Client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
