% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\docType{data}
\name{SparqlDsl}
\alias{SparqlDsl}
\title{SPARQL dsl}
\arguments{
\item{behavior}{Behavior of the error. See Details}
}
\description{
SPARQL dsl
}
\details{
\strong{Methods}
\describe{
 \item{\code{xx}}
}
}
\examples{
\dontrun{
(x <- SparqlDsl$new(url = "http://dbpedia.org/sparql"))
x$url

# ask
x$ask(c("s", "p", "o"))

# select
# SELECT * WHERE { ?s ?p ?o } OFFSET 100 LIMIT 10
x$select$where([:s, :p, :o]).offset(100).limit(10)
}
}
\keyword{datasets}

