% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Peek at a SPARQL query}
\usage{
inspect(.data)
}
\arguments{
\item{.data}{(list) input, using higher level interface}
}
\description{
Peek at a SPARQL query
}
\examples{
sparql_dsl("http://dbpedia.org/sparql") \%>\% ask(s, p, o) \%>\% inspect
}

