% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_exec.R
\name{sparql_exec}
\alias{sparql_exec}
\title{Execute sparql request}
\usage{
sparql_exec(x, query = NULL, ...)
}
\arguments{
\item{x}{a SPARQL server url, or an object of class \code{sparql_dsl}}

\item{query}{(character) A SPARQL query}

\item{...}{Further arguments passed on to \code{\link[httr]{GET}}}
}
\description{
Execute sparql request
}
\examples{
\dontrun{
sparql_exec(spqurl(), "some query")

out <- sparql_exec(spqurl(), query)
head(out)
}
}

